Move diff related classes to own package.
