+++ epoxy-adapter/src/main/java/com/airbnb/epoxy/diff/DiffPayload.java
@@ -1,13 +1,15 @@
-package com.airbnb.epoxy;
+package com.airbnb.epoxy.diff;
 
 import android.support.annotation.Nullable;
 import android.support.annotation.VisibleForTesting;
 import android.support.v4.util.LongSparseArray;
 
+import com.airbnb.epoxy.EpoxyModel;
+
 import java.util.List;
 
 /**
- * A helper class for tracking changed models found by the {@link com.airbnb.epoxy.DiffHelper} to
+ * A helper class for tracking changed models found by the {@link DiffHelper} to
  * be included as a payload in the
  * {@link android.support.v7.widget.RecyclerView.Adapter#notifyItemChanged(int, Object)}
  * call.
