+++ epoxy-adapter/src/main/java/com/airbnb/epoxy/diff/EpoxyDiffLogger.java
@@ -1,4 +1,4 @@
-package com.airbnb.epoxy;
+package com.airbnb.epoxy.diff;
 
 import android.support.annotation.Nullable;
 import android.support.v7.widget.RecyclerView.AdapterDataObserver;
