+++ epoxy-adapter/src/main/java/com/airbnb/epoxy/diff/UpdateOpHelper.java
@@ -1,16 +1,12 @@
-package com.airbnb.epoxy;
+package com.airbnb.epoxy.diff;
 
 import android.support.annotation.Nullable;
 
-import com.airbnb.epoxy.UpdateOp.Type;
+import com.airbnb.epoxy.EpoxyModel;
+import com.airbnb.epoxy.diff.UpdateOp.Type;
 
 import java.util.ArrayList;
 import java.util.List;
-
-import static com.airbnb.epoxy.UpdateOp.ADD;
-import static com.airbnb.epoxy.UpdateOp.MOVE;
-import static com.airbnb.epoxy.UpdateOp.REMOVE;
-import static com.airbnb.epoxy.UpdateOp.UPDATE;
 
 /** Helper class to collect changes in a diff, batching when possible. */
 class UpdateOpHelper {
@@ -45,14 +41,14 @@ class UpdateOpHelper {
 
     // We can append to a previously ADD batch if the new items are added anywhere in the
     // range of the previous batch batch
-    boolean batchWithLast = isLastOp(ADD)
+    boolean batchWithLast = isLastOp(UpdateOp.ADD)
         && (lastOp.contains(startPosition) || lastOp.positionEnd() == startPosition);
 
     if (batchWithLast) {
       addItemsToLastOperation(itemCount, null);
     } else {
       numInsertionBatches++;
-      addNewOperation(ADD, startPosition, itemCount);
+      addNewOperation(UpdateOp.ADD, startPosition, itemCount);
     }
   }
 
@@ -61,7 +57,7 @@ class UpdateOpHelper {
   }
 
   void update(final int indexToChange, EpoxyModel<?> payload) {
-    if (isLastOp(UPDATE)) {
+    if (isLastOp(UpdateOp.UPDATE)) {
       if (lastOp.positionStart == indexToChange + 1) {
         // Change another item at the start of the batch range
         addItemsToLastOperation(1, payload);
@@ -75,10 +71,10 @@ class UpdateOpHelper {
         addItemsToLastOperation(0, payload);
       } else {
         // The item can't be batched with the previous update operation
-        addNewOperation(UPDATE, indexToChange, 1, payload);
+        addNewOperation(UpdateOp.UPDATE, indexToChange, 1, payload);
       }
     } else {
-      addNewOperation(UPDATE, indexToChange, 1, payload);
+      addNewOperation(UpdateOp.UPDATE, indexToChange, 1, payload);
     }
   }
 
@@ -90,7 +86,7 @@ class UpdateOpHelper {
     numRemovals += itemCount;
 
     boolean batchWithLast = false;
-    if (isLastOp(REMOVE)) {
+    if (isLastOp(UpdateOp.REMOVE)) {
       if (lastOp.positionStart == startPosition) {
         // Remove additional items at the end of the batch range
         batchWithLast = true;
@@ -106,7 +102,7 @@ class UpdateOpHelper {
       addItemsToLastOperation(itemCount, null);
     } else {
       numRemovalBatches++;
-      addNewOperation(REMOVE, startPosition, itemCount);
+      addNewOperation(UpdateOp.REMOVE, startPosition, itemCount);
     }
   }
 
@@ -132,7 +128,7 @@ class UpdateOpHelper {
   void move(int from, int to) {
     // We can't batch moves
     lastOp = null;
-    UpdateOp op = UpdateOp.instance(MOVE, from, to, null);
+    UpdateOp op = UpdateOp.instance(UpdateOp.MOVE, from, to, null);
     opList.add(op);
     moves.add(op);
   }
